backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
python-multipart==0.0.6
Pillow==10.1.0
pydantic==1.10.11
